requirements.txt
absolufy-imports==0.3.1
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
Authlib==1.6.4
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
cryptography==46.0.1
docstring_parser==0.17.0
fastapi==0.118.0
google-adk==1.15.1
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.117.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-logging==3.12.1
google-cloud-monitoring==2.27.2
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.58.0
google-cloud-speech==2.33.0
google-cloud-storage==2.18.2
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.39.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langfuse
Mako==1.3.10
pytest==8.3.4
pytest-asyncio==0.24.0
MarkupSafe==3.0.3
mcp==1.15.0
numpy==2.3.3
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.48.0
tenacity==8.5.0
tiktoken==0.8.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
zipp==3.23.0
